Add --extraFieldValue argument
The value of the argument should be a string of the form
`fieldName=value`, and that field/value pair will be added to the
`extra` section of each post which is output. `--extraFieldValue` can be
specified multiple times.

The purpose of this argument is to allow the user to add a constant tag
of some sort to each post which is output, e.g. the name of the blog.