Clean author and title in file output, and allow author to be overwritten
This commit causes output file names to be sanitized to only have
alphanumeric characters and spaces. It's possible for a blog (or its
fucked up RSS feed generator) to add any arbitrary characters to its
title, e.g. newlines (a real life example).

This also allows for overwritting the author field using the
`--extraFieldValue` argument, which is useful if the blog has a missing
or mangled author field.